Create `PhenopacketStoreAuditor` backed by `phenosentry`.
