+++ src/ppktstore/validation/__init__.py
@@ -1,4 +1,7 @@
 from ._impl import qc_phenopacket_store
-from ._checks import UniquePhenopacketId
+from ._ps_auditor import PhenopacketStoreAuditor
 
-__all__ = ["qc_phenopacket_store", "UniquePhenopacketId"]
+__all__ = [
+    "qc_phenopacket_store",
+    "PhenopacketStoreAuditor",
+]
