+++ tests/test_data/README.md
@@ -1,0 +1,3 @@
+# README
+
+We use a real HPO file `v2024-04-26` for testing. When compressed, the file has ~2.1MB which is not too bad™.
